% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compareCounts}
\alias{compareCounts}
\title{Count matrix differences}
\usage{
compareCounts(
  sce1,
  sce2,
  stats = c("pearson", "spearman", "kendall", "mi", "bicor", "dcor")
)
}
\arguments{
\item{sce1}{A \code{SingleCellExperiment}.}

\item{sce2}{A \code{SingleCellExperiment}.}

\item{stats}{Which statistics to compute.}
}
\value{
A \code{data.frame}.
}
\description{
Compute various statistics comparing two count matrices.
}
