% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{cor2vec}
\alias{cor2vec}
\title{Parametrize correlation matrix as unconstrained vector}
\usage{
cor2vec(R)
}
\arguments{
\item{R}{A d x d correlation matrix.}
}
\value{
A vector of length choose(d, 2).
}
\description{
Parametrize correlation matrix as unconstrained vector
}
