% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{dcor_mat}
\alias{dcor_mat}
\title{Pairwise distance correlation matrix}
\usage{
dcor_mat(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix.
}
\description{
Compute matrix of pairwise distance correlations.
}
