% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{empcdf}
\alias{empcdf}
\title{Empirical Cumulative Distribution Function}
\usage{
empcdf(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A function implementing the ECDF of x.
}
\description{
This is simply stats::ecdf scaled by n/(n+1) to avoid
the ECDF evaluating to 1.
}
