% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{fitGaussian}
\alias{fitGaussian}
\title{Fit a Gaussian copula to a SingleCellExperiment}
\usage{
fitGaussian(
  sce,
  margins = c("empirical", "par"),
  estimate = c("sample", "jitter", "mle"),
  cores = 1L
)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{Type of marginal distribution. Either "empirical" or "par". If
the former, margins are modeled empirically. If the latter, margins are
modeled as negative binomial/zero-inflated negative binomial (selected by
AIC).}

\item{estimate}{How should the correlation matrix be estimated. One of
\code{"sample"}, \code{"jitter"}, or \code{"mle"}. The first uses the
sample correlation matrix of the pseudo-observations. The second uses the
sample correlation matrix of jittered pseudo-observations. The third performs
maximum likelihood estimation using the count data.}

\item{cores}{Number of cores.}
}
\value{
A \code{gaussiancop} object.
}
\description{
Fit a Gaussian copula to a SingleCellExperiment
}
