% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{fitIndep}
\alias{fitIndep}
\title{Fit an independence copula to a SingleCellExperiment}
\usage{
fitIndep(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment.}
}
\value{
An \code{indepcop} object.
}
\description{
Fit an independence copula to a SingleCellExperiment
}
