% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{fitT}
\alias{fitT}
\title{Fit a t copula to a SingleCellExperiment}
\usage{
fitT(sce, margins = c("empirical", "par"), Sigma = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{Type of marginal distribution. Either "empirical" or "par". If
the former, margins are modeled empirically. If the latter, margins are
modeled as negative binomial/zero-inflated negative binomial (selected by
AIC).}

\item{Sigma}{Optional precomputed estimate of scale matrix.}
}
\value{
A \code{tcop} object.
}
\description{
Fit a t copula to a SingleCellExperiment
}
