% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vine.R
\name{fitVine}
\alias{fitVine}
\title{Fit a vine copula to a SingleCellExperiment}
\usage{
fitVine(
  sce,
  margins = c("empirical", "par"),
  jitter = FALSE,
  family_set = c("indep", "gaussian", "clayton", "gumbel", "frank", "joe"),
  cores = 1L
)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{Type of marginal distribution. Either "empirical" or "par". If
the former, margins are modeled empirically. If the latter, margins are
modeled as negative binomial/zero-inflated negative binomial (selected by
AIC).}

\item{jitter}{Logical indicating whether data should be jittered.}

\item{family_set}{Family set to use for vine copula construction.}

\item{cores}{Number of cores for parallel computation.}
}
\value{
A vine copula object.
}
\description{
Fit a vine copula to a SingleCellExperiment
}
