% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{gaussian_mle_par}
\alias{gaussian_mle_par}
\title{Maximum Likelihood Estimation of Gaussian Correlation Matrix for Count Data}
\usage{
gaussian_mle_par(FX, FXm, cores)
}
\arguments{
\item{FX}{Margins evaluated at samples.}

\item{FXm}{Margins evaluated at left limits of samples.}

\item{cores}{Number of cores to use for parallel computation.}
}
\value{
Maximum likelihood estimation of correlation matrix.
}
\description{
Maximum likelihood estimation of the correlation matrix for a Gaussian
copula model of count data. Optimization is parallelized.
}
