% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{loglikCopulaCount}
\alias{loglikCopulaCount}
\title{Copula Likelihood for Count Data}
\usage{
loglikCopulaCount(X, copula, margins, cores = 1L)
}
\arguments{
\item{X}{Matrix of data.}

\item{copula}{A \code{copula} object.}

\item{margins}{Vector of marginal distribution functions.}

\item{cores}{Number of cores to use for parallel computation.}
}
\value{
A scalar value.
}
\description{
Compute log-likelihood for a copula model of count data.
}
\details{
This function takes a finite-difference approach and requires 2^d
evaluations of the copula distribution function.
}
