% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{loglikCopulaCts}
\alias{loglikCopulaCts}
\title{Copula Likelihood for Continuous Data}
\usage{
loglikCopulaCts(X, copula, margins = NULL)
}
\arguments{
\item{X}{Matrix of data.}

\item{copula}{A \code{copula} object.}

\item{margins}{A list of marginal distribution functions. If \code{NULL},
samples are first converted to pseudo-observations.}
}
\value{
A scalar value.
}
\description{
Compute log-likelihood for a copula model of continuous data.
}
