% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{loglikEllipCount}
\alias{loglikEllipCount}
\title{Elliptical Copula Likelihood for Count Data}
\usage{
loglikEllipCount(X, copula, margins, cores = 1L)
}
\arguments{
\item{X}{Matrix of data.}

\item{copula}{A \code{normalCopula} or \code{tCopula} object.}

\item{margins}{Vector of marginal distribution functions.}

\item{cores}{Number of cores to use for parallel computation.}
}
\value{
A scalar value.
}
\description{
Compute log-likelihood for an elliptical copula model of count
data.
}
