% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{mi_mat}
\alias{mi_mat}
\title{Pairwise mutual information matrix}
\usage{
mi_mat(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix.
}
\description{
Compute upper triangular matrix of pairwise mutual information.
}
