% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{mut_inf}
\alias{mut_inf}
\title{Pairwise mutual information matrix}
\usage{
mut_inf(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix.
}
\description{
Compute matrix of pairwise mutual information.
}
