% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{nmi_mat}
\alias{nmi_mat}
\title{Pairwise normalized mutual information matrix}
\usage{
nmi_mat(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix.
}
\description{
Compute upper triangular matrix of pairwise normalized mutual
information.
}
