% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{par_margin}
\alias{par_margin}
\title{Parametric CDF for count data}
\usage{
par_margin(x)
}
\arguments{
\item{x}{Numeric vector of count data.}
}
\value{
A function implementing the CDF of x.
}
\description{
Constructs either a negative binomial or zero-inflated negative
binomial CDF for count data. The family is automatically selected by
minimizing AIC.
}
