% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{par_quantile}
\alias{par_quantile}
\title{Parametric quantile functions for count data}
\usage{
par_quantile(x)
}
\arguments{
\item{x}{Numeric vector of count data.}
}
\value{
A function implementing the quantile function of x.
}
\description{
Constructs either a negative binomial or zero-inflated negative
binomial quantile function for count data. The family is automatically
selected by minimizing AIC.
}
