% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{print.gaussiancop}
\alias{print.gaussiancop}
\title{Print method for S3 class gaussiancop}
\usage{
\method{print}{gaussiancop}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gaussiancop}.}

\item{...}{Additional arguments.}
}
\description{
Print method for S3 class gaussiancop
}
