% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{print.indepcop}
\alias{print.indepcop}
\title{Print method for S3 class indepcop}
\usage{
\method{print}{indepcop}(x, ...)
}
\arguments{
\item{x}{Object of class \code{indepcop}.}

\item{...}{Additional arguments.}
}
\description{
Print method for S3 class indepcop
}
