% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{print.tcop}
\alias{print.tcop}
\title{Print method for S3 class tcop}
\usage{
\method{print}{tcop}(x, ...)
}
\arguments{
\item{x}{Object of class \code{tcop}.}

\item{...}{Additional arguments.}
}
\description{
Print method for S3 class tcop
}
