% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{sampleCells}
\alias{sampleCells}
\title{Sample cells}
\usage{
sampleCells(N, Qx, copula)
}
\arguments{
\item{N}{Number of cells to sample.}

\item{Qx}{List of marginal quantile functions.}

\item{copula}{A copula object. Currently supported are \code{vinecop},
\code{gaussiancop}, \code{tcop}, and \code{indepcop}.}
}
\value{
A SingleCellExperiment.
}
\description{
Sample cells from the model specified by the given copula and
marginal quantile functions.
}
