% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{t_mle}
\alias{t_mle}
\title{Maximum Likelihood Estimation of t Copula Parameters}
\usage{
t_mle(FX, FXm, Sigma = NULL, ...)
}
\arguments{
\item{FX}{Margins evaluated at samples.}

\item{FXm}{Margins evaluated at left limits of samples.}

\item{Sigma}{Optional precomputed estimate of scale matrix.}

\item{...}{Additional arguments to pass to \code{stats::optim}.}
}
\value{
Maximum likelihood estimation of scale matrix and degrees of
freedom.
}
\description{
Maximum likelihood estimation of the scale matrix and degrees of freedom for
a t copula model of count data.
}
