% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{vec2cor}
\alias{vec2cor}
\title{Inverse of cor2vec}
\usage{
vec2cor(v)
}
\arguments{
\item{v}{A vector of length choose(d, 2).}
}
\value{
A d x d correlation matrix.
}
\description{
Inverse of cor2vec
}
