% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compareCounts}
\alias{compareCounts}
\title{Count Matrix Differences}
\usage{
compareCounts(
  sce1,
  sce2,
  assay = "counts",
  stats = c("pearson", "spearman", "kendall", "mi", "bicor", "dcor", "rhoA")
)
}
\arguments{
\item{sce1}{A SingleCellExperiment.}

\item{sce2}{A SingleCellExperiment.}

\item{assay}{The assay to use.}

\item{stats}{Which statistics to compute.}
}
\value{
A vector of statistics.
}
\description{
Compute various statistics comparing two count matrices.
}
