% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{fitGaussian}
\alias{fitGaussian}
\title{Fit a Gaussian copula to a SingleCellExperiment}
\usage{
fitGaussian(
  sce,
  margins,
  mle = TRUE,
  jitter = FALSE,
  likelihood = TRUE,
  cores = 1L
)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{One, or a list of, marginal distribution functions. If only
one function is passed, all margins are modeled identically. Entries can also
equal "empirical" or "par", in which case the corresponding margin is modeled
empirically or as a negative binomial/zero-inflated negative binomial
(selected by AIC), respectively.}

\item{mle}{Logical indicating whether maximum likelihood estimation should
be performed to estimate the correlation matrix. If \code{FALSE}, the sample
correlation matrix is used.}

\item{jitter}{Logical indicating whether data should be jittered.}

\item{likelihood}{Whether to compute likelihood.}

\item{cores}{Number of cores.}
}
\value{
A \code{gaussiancop} object.
}
\description{
Fit a Gaussian copula to a SingleCellExperiment
}
