% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{fitGaussian}
\alias{fitGaussian}
\title{Fit a Gaussian copula to a SingleCellExperiment}
\usage{
fitGaussian(
  sce,
  margins = c("empirical", "par"),
  mle = TRUE,
  jitter = FALSE,
  likelihood = TRUE,
  cores = 1L
)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{Type of marginal distribution. Either "empirical" or "par". If
the former, margins are modeled empirically. If the latter, margins are
modeled as negative binomial/zero-inflated negative binomial (selected by
AIC).}

\item{mle}{Logical indicating whether maximum likelihood estimation should
be performed to estimate the correlation matrix. If \code{FALSE}, the sample
correlation matrix is used.}

\item{jitter}{Logical indicating whether data should be jittered.}

\item{likelihood}{Whether to compute likelihood.}

\item{cores}{Number of cores.}
}
\value{
A \code{gaussiancop} object.
}
\description{
Fit a Gaussian copula to a SingleCellExperiment
}
