% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{fitT}
\alias{fitT}
\title{Fit a t copula to a SingleCellExperiment}
\usage{
fitT(sce, margins, Sigma = NULL, likelihood = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{One, or a list of, marginal distribution functions. If only
one function is passed, all margins are modeled identically. Entries can also
equal "empirical" or "nb", in which case the corresponding margin is modeled
empirically or as a negative binomial, respectively.}

\item{Sigma}{Optional precomputed estimate of scale matrix.}

\item{likelihood}{Whether to compute likelihood.}
}
\value{
A \code{tcop} object.
}
\description{
Fit a t copula to a SingleCellExperiment
}
