% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vine.R
\name{fitVine}
\alias{fitVine}
\title{Fit a vine copula to a SingleCellExperiment}
\usage{
fitVine(
  sce,
  margins,
  jitter = FALSE,
  family_set = c("indep", "gaussian", "clayton", "gumbel", "frank", "joe"),
  cores = 1L
)
}
\arguments{
\item{sce}{A SingleCellExperiment.}

\item{margins}{One, or a list of, marginal distribution functions. If only
one function is passed, all margins are modeled identically. Entries can also
equal "empirical", in which case the corresponding margin is modeled
empirically.}

\item{jitter}{Logical indicating whether data should be jittered.}

\item{family_set}{Family set to use for vine copula construction.}

\item{cores}{Number of cores for parallel computation.}
}
\value{
A vine copula object.
}
\description{
Fit a vine copula to a SingleCellExperiment
}
