% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{gaussian_mle}
\alias{gaussian_mle}
\title{Maximum Likelihood Estimation of Gaussian Correlation Matrix for Count Data}
\usage{
gaussian_mle(FX, FXm, ...)
}
\arguments{
\item{FX}{Margins evaluated at samples.}

\item{FXm}{Margins evaluated at left limits of samples.}

\item{...}{Additional arguments to pass to \code{stats::optim}.}
}
\value{
Maximum likelihood estimation of correlation matrix.
}
\description{
Maximum likelihood estimation of the correlation matrix for a Gaussian
copula model of count data.
}
